% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{sum_func}
\alias{sum_func}
\title{sum_func
A function to sum up the list of number}
\usage{
sum_func(numbs)
}
\arguments{
\item{numbs}{numbs is a list of numbers which will be summed up in the function}
}
\value{
It will return a number which is the sum up of the input
}
\description{
sum_func
A function to sum up the list of number
}
\examples{
sum_func(1:10)

}
