% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression
Get the linear regression parameters}
\usage{
linear_regression(X, Y)
}
\arguments{
\item{X}{predictors values}

\item{Y}{response values

return the least squares estimates in matrix notation}
}
\description{
linear_regression
Get the linear regression parameters
}
\examples{
linear_regression(X = matrix(rnorm(3*100), nrow = 100, ncol = 3), Y = matrix(rnorm(1*100), nrow = 100, ncol = 1))

}
