% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression
Fit the linear models and return the coefficients estimation information}
\usage{
linear_regression(X, Y)
}
\arguments{
\item{X}{predictors input. The X should be the predictors that you want to use to fit the model. It should be either a matrix or a dataframe.}

\item{Y}{response values. The Y should be the response used to fit the linear model. It should be be either a matrix or a dataframe}
}
\value{
The return value is a table with four columns. The first column is the coefficients estimate,  the second columns is the coefficients corresponding standard error, the third columns is the coefficients corresponding t values and the fourth column is the coefficient corresponding p values.
}
\description{
linear_regression
Fit the linear models and return the coefficients estimation information
}
\examples{
linear_regression(X = matrix(rnorm(3*100), nrow = 100, ncol = 3), Y = matrix(rnorm(1*100), nrow = 100, ncol = 1))

}
