% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression
Fit the linear models}
\usage{
linear_regression(X, Y)
}
\arguments{
\item{X}{predictors input. The X should be the predictors that you want to use to fit the model. It should be either a matrix or a dataframe.}

\item{Y}{response values. The Y should be the response used to fit the linear model. It should be be either a matrix or a dataframe}
}
\value{
It will return a list with the linear model coefficients, coefficients corresponding standard error, coefficients t values and coefficients p values
}
\description{
linear_regression
Fit the linear models
}
\examples{
linear_regression(X = matrix(rnorm(3*100), nrow = 100, ncol = 3), Y = matrix(rnorm(1*100), nrow = 100, ncol = 1))

}
